refactor: extract login core CSS to styles folder
