+++ packages/login/src/styles/vaadin-login-form-wrapper-styles.js
